Move utility functions into `lib`
