+++ src/javascript/lib/stringFunctions.jsx
