+++ src/javascript/models/transformations/lineitem/RemoveRepetitiveElements.jsx
@@ -2,7 +2,7 @@ import ToLineItemTransformation from '../ToLineItemTransformation.jsx';
 import ParseResult from '../../ParseResult.jsx';
 import { REMOVED_ANNOTATION } from '../../Annotation.jsx';
 
-import { isDigit } from '../../../stringFunctions.jsx'
+import { isDigit } from '../../../lib/stringFunctions.jsx'
 
 
 function hashCodeIgnoringSpacesAndNumbers(string) {
